requirements.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap-pagination==1.5.0
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
djangorestframework==3.0.0
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
trueskill==0.4.3
